Load Acct Numbers from KeePass; Set default trading account using trading.yaml defaults.account string which matches KeePass entry for account number. Do NOT change GUI Account selection when loading a new template or executing a hotkey/button trade: USER must select which account to use.

+++ src/trading_app/trading_config.py
@@ -290,7 +290,7 @@ class TradingConfig:
                         value["quantity"]["type"].lower()
                     ),
 
-                    value=int(
+                    value=float(
                         value["quantity"]["value"]
                     ),
                 ),
